+++ src/entities/searchbar/search-button/search-button.tsx
@@ -1,13 +1,14 @@
-import { FunctionComponent } from 'react'
-import { Search } from '@mui/icons-material'
-import { BasicButton } from '../../../shared/ui/basic-button'
-import { BaseButton } from '../../../shared/config/types'
+import { FunctionComponent } from "react";
+import { Search } from "@mui/icons-material";
+import { useUIComponents } from "../../../shared/hooks/use-ui-components";
+import { BaseButton } from "../../../shared/config/types";
 
 const SearchButton: FunctionComponent<BaseButton> = ({ onHandler }) => {
+  const { BasicButton } = useUIComponents();
   return (
     <BasicButton onClick={onHandler}>
       <Search />
     </BasicButton>
-  )
-}
-export default SearchButton
+  );
+};
+export default SearchButton;
