+++ src/entities/sidebar/menu-button/menu-button.tsx
@@ -1,13 +1,14 @@
-import { FunctionComponent } from 'react'
-import { Menu } from '@mui/icons-material'
-import { BasicButton } from '../../../shared/ui/basic-button'
-import { BaseButton } from '../../../shared/config/types'
+import { FunctionComponent } from "react";
+import { Menu } from "@mui/icons-material";
+import { BaseButton } from "../../../shared/config/types";
+import { useUIComponents } from "../../../shared/hooks/use-ui-components";
 
 const MenuButton: FunctionComponent<BaseButton> = ({ onHandler }) => {
+  const { BasicButton } = useUIComponents();
   return (
     <BasicButton disableRipple onClick={onHandler}>
       <Menu />
     </BasicButton>
-  )
-}
-export default MenuButton
+  );
+};
+export default MenuButton;
