+++ src/entities/userbar/login-button/login-button.tsx
@@ -1,15 +1,17 @@
-import { FunctionComponent, MouseEvent } from 'react'
-import { AccountCircle } from '@mui/icons-material'
-import { BasicButton } from '../../../shared/ui/basic-button'
+import { FunctionComponent, MouseEvent } from "react";
+import { AccountCircle } from "@mui/icons-material";
+import { useUIComponents } from "../../../shared/hooks/use-ui-components";
 
 interface LoginBtn {
-  onHandler: (e: MouseEvent<HTMLElement>) => void
+  onHandler: (e: MouseEvent<HTMLElement>) => void;
 }
 const LoginButton: FunctionComponent<LoginBtn> = ({ onHandler }) => {
+  const { BasicButton } = useUIComponents();
+
   return (
     <BasicButton onClick={onHandler}>
-      <AccountCircle fontSize='large' />
+      <AccountCircle fontSize="large" />
     </BasicButton>
-  )
-}
-export default LoginButton
+  );
+};
+export default LoginButton;
