+++ src/shared/hooks/use-ui-components.ts
@@ -1,0 +1,7 @@
+import { createContext, useContext } from 'react'
+import UIComponents from '../ui'
+
+export const UIContext = createContext(UIComponents)
+export const useUIComponents = () => {
+  return useContext(UIContext)
+}
