+++ src/shared/ui/index.ts
@@ -1,0 +1,2 @@
+import { BasicButton, CatButton, CatLink } from '../ui/buttons'
+export default { BasicButton, CatButton, CatLink }
