+++ src/shared/ui/main-logo/main-logo.tsx
@@ -1,84 +1,12 @@
-import s from './config/logo.module.css';
-const MainLogo = () => {
-  return (
-    <a style={{marginLeft: '30px'}} href="/">
-        <svg version="1.0" 
-         width="30pt" height="30pt" viewBox="0 0 512.000000 512.000000"
-         preserveAspectRatio="xMidYMid meet">
-        
-        <g transform="translate(0.000000,512.000000) scale(0.100000,-0.100000)"
-        fill="#000000" stroke="none">
-        <path d="M791 5106 c-50 -27 -50 -34 -51 -453 l0 -393 -230 0 c-251 0 -275 -4
-        -298 -55 -9 -20 -12 -481 -12 -1954 0 -1903 0 -1930 20 -1992 37 -119 123
-        -203 242 -241 57 -17 144 -18 2098 -18 2012 0 2039 0 2101 20 119 37 203 123
-        241 242 17 56 18 107 18 764 0 765 1 747 -55 772 -33 15 -57 15 -90 0 -56 -25
-        -55 -8 -55 -763 0 -568 -3 -702 -14 -728 -16 -39 -61 -82 -99 -96 -20 -8 -592
-        -11 -1867 -11 l-1838 0 19 61 c18 59 19 131 19 2360 l0 2299 1890 0 1890 0 0
-        -1165 c0 -816 3 -1171 11 -1188 12 -27 58 -57 88 -57 33 0 67 20 84 50 16 26
-        17 128 17 1255 0 1332 3 1267 -55 1293 -40 19 -4040 17 -4074 -2z m-51 -2906
-        c0 -1540 -2 -1866 -14 -1893 -30 -72 -114 -117 -191 -103 -43 8 -103 59 -121
-        103 -12 27 -14 353 -14 1893 l0 1860 170 0 170 0 0 -1860z"/>
-        <path d="M1254 4500 c-12 -4 -31 -21 -43 -36 -21 -26 -21 -34 -21 -391 0 -395
-        2 -407 55 -431 14 -7 35 -12 46 -12 28 0 75 26 85 47 5 10 11 105 14 212 l5
-        193 138 -201 c76 -110 150 -210 164 -221 34 -26 92 -26 126 1 45 35 48 68 45
-        442 l-3 349 -33 29 c-18 16 -44 29 -58 29 -33 0 -81 -25 -94 -49 -6 -12 -10
-        -100 -10 -216 0 -109 -4 -194 -9 -191 -4 3 -71 96 -147 207 -76 111 -152 212
-        -167 225 -30 25 -57 29 -93 14z"/>
-        <path d="M2118 4495 c-15 -9 -32 -24 -38 -35 -6 -12 -10 -157 -10 -394 l0
-        -375 24 -28 24 -28 192 0 192 0 24 28 c33 38 33 96 0 134 -24 27 -27 28 -140
-        31 l-116 4 0 69 0 69 90 0 c128 0 170 25 170 100 0 75 -42 100 -170 100 l-90
-        0 0 69 0 69 116 4 c113 3 116 4 140 31 33 38 33 96 0 134 l-24 28 -179 2
-        c-144 2 -183 0 -205 -12z"/>
-        <path d="M2772 4485 c-16 -14 -33 -36 -36 -49 -3 -14 24 -178 64 -383 l70
-        -358 35 -32 c25 -23 44 -33 67 -33 37 0 92 30 106 57 5 10 33 98 63 195 29 98
-        57 172 60 165 4 -6 30 -88 59 -181 57 -188 62 -197 101 -220 52 -31 126 -11
-        153 42 17 33 146 674 146 726 0 63 -66 109 -127 89 -48 -16 -66 -57 -97 -226
-        -16 -87 -32 -160 -36 -162 -3 -2 -30 75 -60 171 -45 146 -58 179 -82 199 -35
-        30 -72 32 -114 6 -26 -16 -35 -33 -57 -102 -14 -46 -40 -131 -58 -189 l-33
-        -105 -31 160 c-43 218 -41 213 -78 235 -43 27 -79 25 -115 -5z"/>
-        <path d="M4075 4501 c-85 -24 -156 -79 -186 -146 -26 -56 -25 -147 2 -202 30
-        -62 75 -94 209 -147 171 -68 199 -96 146 -145 -52 -49 -148 -37 -236 30 -49
-        36 -60 40 -95 35 -53 -6 -85 -43 -85 -96 0 -32 7 -47 38 -78 97 -101 259 -146
-        389 -107 76 23 142 72 175 131 57 103 47 199 -30 285 -49 55 -86 76 -234 134
-        -83 32 -98 41 -98 60 0 62 117 72 197 18 21 -14 54 -28 74 -30 30 -4 42 0 70
-        25 28 25 34 37 34 71 0 53 -35 89 -126 131 -74 34 -183 48 -244 31z"/>
-        <path d="M1241 3206 c-52 -29 -50 11 -51 -1043 0 -686 3 -989 11 -1006 6 -13
-        23 -31 39 -40 26 -16 104 -17 889 -17 597 0 867 3 884 11 13 6 31 23 40 39 16
-        26 17 112 17 1010 0 1066 2 1022 -55 1048 -40 18 -1741 16 -1774 -2z m1629
-        -468 l0 -282 -158 157 c-156 155 -159 157 -202 157 l-44 0 -251 -250 -250
-        -250 -105 106 c-173 174 -175 174 -347 1 l-123 -122 0 383 0 382 740 0 740 0
-        0 -282z m-178 -385 l178 -178 0 -405 0 -405 -382 382 -383 383 200 200 c110
-        110 202 200 205 200 3 0 85 -80 182 -177z m-517 -573 l480 -480 -633 0 -632 0
-        0 337 0 338 142 142 c79 79 147 143 153 143 6 0 226 -216 490 -480z"/>
-        <path d="M3421 3206 c-66 -37 -67 -131 -1 -169 25 -15 78 -17 499 -17 317 0
-        478 4 494 11 28 13 57 58 57 89 0 33 -26 75 -55 88 -40 18 -961 16 -994 -2z"/>
-        <path d="M3421 2726 c-66 -37 -67 -131 -1 -169 25 -15 78 -17 499 -17 317 0
-        478 4 494 11 28 13 57 58 57 89 0 33 -26 75 -55 88 -40 18 -961 16 -994 -2z"/>
-        <path d="M3421 2246 c-66 -37 -67 -131 -1 -169 25 -15 78 -17 499 -17 317 0
-        478 4 494 11 28 13 57 58 57 89 0 33 -26 75 -55 88 -40 18 -961 16 -994 -2z"/>
-        <path d="M4771 2246 c-66 -37 -67 -131 -1 -169 87 -52 180 35 138 128 -22 48
-        -88 68 -137 41z"/>
-        <path d="M3421 1766 c-66 -37 -67 -131 -1 -169 25 -15 78 -17 499 -17 317 0
-        478 4 494 11 28 13 57 58 57 89 0 33 -26 75 -55 88 -40 18 -961 16 -994 -2z"/>
-        <path d="M3421 1286 c-66 -37 -67 -131 -1 -169 25 -15 78 -17 499 -17 317 0
-        478 4 494 11 28 13 57 58 57 89 0 33 -26 75 -55 88 -40 18 -961 16 -994 -2z"/>
-        <path d="M1241 726 c-67 -37 -65 -143 4 -174 20 -9 398 -12 1586 -12 1331 0
-        1564 2 1588 14 66 35 66 137 0 172 -40 20 -3141 21 -3178 0z"/>
-        </g>
-        </svg>
-    </a>
-  )
-}
+import s from "./config/logo.module.css";
 
 const Logo = () => {
-
   return (
-    <a href='/'>
-      <div className={s['logo-module']}>
+    <a href="/">
+      <div className={s["logo-module"]}>
         <span>NewsRoom</span>
       </div>
     </a>
-
-  )
-}
+  );
+};
 export default Logo;
